Add support for Android 11 (minSdkVersion 23)
This commit updates the Android build configuration to set the minSdkVersion to 23, which is required for Android 11 support. Additionally, it renames the app name from "Ennovation" to "E-Mart" in the app_strings.dart file.

The commit also updates the dependencies to their latest versions, including:
- velocity_x: ^4.2.1
- firebase_core: ^3.1.0
- firebase_auth: ^5.1.0
- firebase_storage: ^12.0.1
- cloud_firestore: ^5.0.1
- image_picker: ^1.1.2
- internet_connection_checker: ^2.0.0
- intl: ^0.19.0
- flutter_launcher_icons: ^0.13.1
- flutter_lints: ^4.0.0

The updated dependencies bring in bug fixes, performance improvements, and new features to the project.